Improve the way courses are retrieved
Currently, 3 separate qureries are fired to retrive ongoing, upcoming
and finished courses. Sometimes some of the courses slip through the
cracks and disappear from the front page.

To fix this, just use a single query and then filter the courses in
code. Also, add another status value to the offering table status field
which represents 'Course not available'. These courses are to be
filtered out from the front pages.