Add column short_name to offering table
This column will be used to uniquely identify each offering. This is
mainly done to parse the new batch of coursera courses. For the coursera
courses the shortname will be as follows:

(coursera_short_name)_(course_id)_(offering_id)

All the for short name details are taken from the courses JSON at:

https://www.coursera.org/maestro/api/topic/list?full=1